technical/server/api/requirements.txt
flask
paho-mqtt
psycopg2-binary
redis
gunicorn